Replace `standard` with `prettier`
It looks like `prettier` replaced `standard` as the most popular
JavaScript style in 2017 or so:

https://www.reddit.com/r/javascript/comments/6g8adf/prettier_has_surpassed_standard_as_mostdownloaded/

In general, my personal preference is to use a lagnuage's most popular
coding style and to "enforce" it by using an autoformatter instead of a
manual linter. This means I can just hit "save" in my editor and I'll
follow all the rules—no manually fixing lint errors.

I agree that this is a reckless and excessive thing to do. But I'm about
to convert everything to TypeScript anyway, so let's be reckless.

We also fix a unit test that broke because of changing error messages.
And we take tests out of the precommit hook because they require db
access, and that's hard to set up inside some editor's git integration.

We can always restore tests using GitHub actions at some point.